% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiPOST.R
\name{apiPOST}
\alias{apiPOST}
\title{apiPOST - Do a POST to the Core ODATA REST API.}
\usage{
apiPOST(coreApi, resource = NULL, body = NULL, encode,
  headers = NULL, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{entity type for POST}

\item{body}{body for request}

\item{encode}{encode type must be "multipart", "form", "json", "raw"}

\item{headers}{headers to be added to get.}

\item{special}{passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
Returns the entire http response
}
\description{
\code{apiPOST} - Base call to Core ODATA REST API.
}
\details{
\code{apiPOST}  Do a POST to the Core ODATA REST API.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- apiPOST(login$coreApi, "SAMPLE", body, "json", special = NULL, useVerbose = FALSE)
message <- httr::content(response)
error <- httr::http_error(response)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Francisco Marin francisco.marin@thermofisher.com
}
