% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCleanName.R
\name{attributeCleanName}
\alias{attributeCleanName}
\title{attributeCleanName - converts names to uppercase and replaces spaces
ampersands and percent symbols with underscores to reference attribute
names.
\code{attributeCleanName} clean attribute name.}
\usage{
attributeCleanName(name)
}
\arguments{
\item{name}{string to clean}
}
\value{
Returns attribute name as the API will use.
}
\description{
\code{attriureCleanName} - onverts names to uppercase and replaces spaces with underscores to
reference attribute names.
}
\examples{
\dontrun{
new_name <- attributeCleanName("My Comments")
new_name
MY_COMMENTS
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com
}
