% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{case}
\alias{case}
\title{case - Performs a vectorized IF}
\usage{
case(...)
}
\arguments{
\item{...}{A set of two sided formulas as LHS~RHS where LHS is logical and RHS is any closure}
}
\value{
Nothing is returned the first case to evaluate as true will execute in the parent environment
This function is purely used for its side effects.
}
\description{
\code{case} Performs a vectorized IF
}
\details{
\code{case} Performs a vectorized IF
}
\examples{
\dontrun{
case(
   grepl("[0-2]+\\\\.[0-9]+\\\\.[0-9]+","3.7.1") ~ {
     body <- list("a",
                  "b")
   },
   grepl("[3-9]+\\\\.[0-9]+\\\\.[0-9]+","3.7.1") ~ {
     body <- list("b",
                  "a")
   } 
)
}
}
\author{
Adam Wheeler, adam.j.wheeler@accenture.com
}
