% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{coreAPI - Creates a object of class coreAPI that contains user and connection information.}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with tenant information in json format.}
}
\value{
Object of class coreAPI
}
\description{
\code{coreAPI} Creates a object of class coreAPI that contains user and connection information.
}
\details{
{ Creates a object of class list that contains connection information.
         Requires a json formatted datastring.
        \code{#'Creates a object of class list that contains connection information}
        \code{coreAPI("path to json")}.
}
The json must include the fields shown below.
    \code{
{
  "values": [
    {
      "key": "tenant",
      "value": "FULLNAMEOFTENANT",
      "type": "text",
      "enabled": true
    },
    {
      "key": "alias",
      "value": "SHORTNAMEINURL",
      "type": "text",
      "enabled": true
    },
    {
      "key": "scheme",
      "value": "https",
      "type": "text",
      "name": "scheme",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "host",
      "value": "HOSTNAME",
      "type": "text",
      "name": "host",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "port",
      "value": "443",
      "type": "text",
      "enabled": true
    },
    {
      "key": "context",
      "value": "",
      "type": "text",
      "enabled": true
    },
    {
      "key": "api_username",
      "value": "USER",
      "type": "text",
      "name": "nameadmin",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "api_password",
      "value": "PASSWORD",
      "type": "text",
      "name": "passwordadmin",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "semver",
      "value": "",
      "type": "text",
      "enabled": true
    }
    ]
}
    }

 The tenant value may be set to "" if the user only has access to one tenant.
As an alternative the environment json object from Postman can be used directly.
}
\examples{
\dontrun{
api <- coreAPI("/home/environment.json")
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Scott Russell scott.russell@thermofisher.com

Adam Wheeler, adam.j.wheeler@accenture.com

Natasha Mora natasha.mora@thermofisher.com
}
