% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperiment.R
\name{createExperiment}
\alias{createExperiment}
\title{createExperiment - Create a new instance of an experiment.}
\usage{
createExperiment(coreApi, experimentType, assayType, assayBarcode,
  protocolType, protocolBarcode, body = NULL, fullMetadata = FALSE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment type to get as character string}

\item{assayType}{assay type}

\item{assayBarcode}{assay barcode}

\item{protocolType}{protocol type}

\item{protocolBarcode}{protocol barcode}

\item{body}{values for experiment attributes and associations as a  list of key-values pairs}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createExperiment} Creates a new experiment.
}
\details{
\code{createExperiment} Creates a new instance of an entity.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experiment <- createExperiment(
  login$coreApi,
  "Experiment_Type",
  "Assaybarcode",
  "Protocolbarcode"
)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
