% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentPublish.R
\name{experimentPublish}
\alias{experimentPublish}
\title{experimentPublish Publishes an experiment.}
\usage{
experimentPublish(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment (Experiments that require a signature cannot be published through the API).}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information, $response contains the entire http response
For PFS version 5.x.x the output is generate as list(entity = httr::content(response)$response$data, response = response)
For PFS version 6.x.x the output is generate as list(entity = httr::content(response), response = response)
}
\description{
\code{experimentPublish} - Publishes an experiment.
}
\details{
\code{experimentPublish} Publishes an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
update <- experimentPublish(login$coreApi, experimentType, experimentBarcode, useVerbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Natasha Mora, natasha.mora@thermofisher.com
}
