% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentUnpublish.R
\name{experimentUnpublish}
\alias{experimentUnpublish}
\title{experimentUnpublish Unpublishes an experiment.}
\usage{
experimentUnpublish(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information, $response contains the entire http response
}
\description{
\code{experimentUnpublish} - Unpublishes an experiment.
}
\details{
\code{experimentUnpublish} Unpublishes an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
update <- experimentUnpublish(login$coreApi, experimentType, experimentBarcode, useVerbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
