% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{getAccountInfoValue}
\alias{getAccountInfoValue}
\title{getAccountInfoValue - read a configuration value from a data frame or environment variable}
\usage{
getAccountInfoValue(accountinfo, key)
}
\arguments{
\item{accountinfo}{data frame of configuration values}

\item{key}{name of data frame key that might match an environment variable}
}
\value{
value from the data frame or environment variable
}
\description{
getAccountInfoValue - read a configuration value from a data frame or environment variable
}
\author{
Scott Russell scott.russell@thermofisher.com
}
