% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttachedAttributeFile.R
\name{getAttachedFile-deprecated}
\alias{getAttachedFile-deprecated}
\title{getAttachedFile - Gets file attached to an attribute on a entity}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type where sample is attached}

\item{barcode}{barcode of the entity}

\item{attribute}{name of the attribute}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used}
}
\value{
returns a list $entity contains file data, $response contains the entire http response
}
\description{
\code{ getAttachedFile }  Gets file attached to an attribute on a entity.
The $entity slot contains the binary file data.
}
\details{
\code{getAttachedFile}  Gets file attached to an attribute on a entity
}
\examples{
\dontrun{
api <- CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login <- CoreAPIV2::authBasic(api)
response <- getAttachedFile(login$coreApi, entityType, barcode, attribute)
witeBin(response$entity, "filename.txt")
CoreAPIV2:logOut(login$coreApi)
}
}
\seealso{
\code{\link{CoreAPIV2-deprecated}}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
\keyword{internal}
