% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellContents.R
\name{getCellContents}
\alias{getCellContents}
\title{getCellContents -  Gets information about a single container cell contents.}
\usage{
getCellContents(coreApi, containerCellId, fullMetadata = FALSE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerCellId}{container cell number as a string}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getCellContents} -  Gets information about a single container cell contents.
}
\details{
\code{getCellContents} Gets information about container cell contents.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
cell <- getCellContents(login$coreApi, "1234233", fullMetadata = TRUE)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
