% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContainerCellIds.R
\name{getContainerCellIds}
\alias{getContainerCellIds}
\title{getContainerCellIds -  Gets cell ids for a container}
\usage{
getContainerCellIds(coreApi, containerBarcode,
  containerType = "CONTAINER", useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{container barcode}

\item{containerType}{container entity type (default: CONTAINER)}

\item{useVerbose}{Use verbose communication for debugging (default: FALSE)}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is a vector of cell IDs.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getContainerCellIds} -  Gets cell ids for a container
}
\details{
\code{getContainerCellIds} Gets information about container contents.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
cellIDs <- getContainerCellIds(login$coreApi, "TE1", containerType = "384 WELL PLATE")$entity
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
