% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContainerContents.R
\name{getContainerContents}
\alias{getContainerContents}
\title{getContainerContents -  Gets container contents.}
\usage{
getContainerContents(coreApi, containerBarcode,
  containerType = "CONTAINER", fullMetadata = TRUE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{container barcode}

\item{containerType}{container entity type}

\item{fullMetadata}{return full metadata}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getContainerContents} - Gets information about container cell contents.
}
\details{
\code{getContainerContents} Gets information about container contents.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
cell <- getContainerContents(login$coreApi, "VIA9", "VIAL")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
