% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityAssociations.R
\name{getEntityAssociations}
\alias{getEntityAssociations}
\title{\code{getEntityAssociations} Get associations for an entity}
\usage{
getEntityAssociations(coreApi, entityType, barcode, associationContext,
  fullMetadata = FALSE, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{associationContext}{association context}

\item{fullMetadata}{- get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of entity associations
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getEntityAssociations}  Get associations for an entity
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
associations <- getEntityAssociations(login$coreAPI, "entityType", "barcode", "associationContext")
logOut(login$coreAPI)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
