% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityAttributeMetadata.R
\name{getEntityAttributeMetadata}
\alias{getEntityAttributeMetadata}
\title{getEntityAttributeMetadata - Gets attribute metadata for a particular entity type.
\code{getEntityAttributeMetadata} Gets attribute metadata for a particular entity type.}
\usage{
getEntityAttributeMetadata(coreApi, entityType, fullMetadata = FALSE,
  ...)
}
\arguments{
\item{entityType}{entity type which has the attribute metadata}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}

\item{coreAPI}{coreApi object with valid jsessionid}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of attribute metadata for the specified entity.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getEntityAttributeMetadata} - Gets attribute metadata for a particular entity type.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experimentAssayType <- getEntityAttributeMetadata(login$coreApi, "BEER_SAMPLE", FALSE)
logOut(login$coreApi)
}
}
\author{
Natasha Mora natasha.mora@thermofisher.com

Francisco Marin francisco.marin@thermofisher.com

Edgardo Gutierrez edgardo.gutierrez@thermofisher.com
}
