% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityByBarcode.R
\name{getEntityByBarcode}
\alias{getEntityByBarcode}
\title{getEntityByBarcode - Get an entity by barcode from PFS using the OData API.}
\usage{
getEntityByBarcode(coreApi, entityType, barcode, fullMetadata = FALSE,
  ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{fullMetadata}{- get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content.
 \item{\code{response}} NULL by default (parameter \code{fullReturn} is TRUE).
  If \code{fullReturn} is FALSE, returns the entire HTTP response.
}
}
\description{
\code{getEntityByBarcode} Get an entity by barcode from PFS using the OData API.
}
\details{
\code{getEntityByBarcode} Get an entity by barcode from PFS using the OData API.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
item <- getEntityByBarcode(login$coreApi, "entityType", "barcode")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
