% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityMetadata.R
\name{getEntityMetadata}
\alias{getEntityMetadata}
\title{getEntityMetadata - Get entity metadata.}
\usage{
getEntityMetadata(coreApi, entityType, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get metadata for}

\item{...}{additional arguments passed to \code{apiGET}
#' @return List of length 3, containing \code{attributes}, \code{associations}, and \code{template} objects:
\itemize{
 \item{\code{attributes}} is a list of attribute properties.
 \item{\code{response}} is a list of association properties.
 \item{\code{template}} is a list convertable to JSON for object creation.
}}
}
\description{
\code{getEntityMetadata} Get entity metadata by entityType. Returns a list with three data frames
named attributes, associations, and template. Template can be used as the body for createEntity functions.
}
\details{
\code{getEntityMetadata}  Get entity metadata.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
item <- getEntityMetadata(login$coreApi, "entityType")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
