% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainerCellIds.R
\name{getExperimentContainerCellIds}
\alias{getExperimentContainerCellIds}
\title{getExperimentContainerCellIds - Gets cell ids for a container in an experiment.}
\usage{
getExperimentContainerCellIds(coreApi, experimentContainerBarcode,
  experimentContainerType = "EXPERIMENT_CONTAINER", useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentContainerBarcode}{container barcode}

\item{experimentContainerType}{container entity type (default: EXPERIMENT_CONTAINER)}

\item{useVerbose}{Use verbose communication for debugging (default: FALSE)}
}
\value{
RETURN returns $entity a array of cell IDs and  $response contains the entire http response
}
\description{
\code{getExperimentContainerCellIds} - Gets cell ids for a container in an experiment.
}
\details{
\code{getExperimentContainerCellIds} Gets cell ids for a container in an experiment.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
cellIDs<-CoreAPIV2::getExperimentContainerCellIds(login$coreApi, "BTCR1", "BITTERNESS EXPERIMENT CONTAINER")$entity
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Scott Russell scott.russell@thermofisher.com
}
