% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainerContents.R
\name{getExperimentContainerContents}
\alias{getExperimentContainerContents}
\title{getExperimentContainerContents -  Gets experiment container contents.}
\usage{
getExperimentContainerContents(coreApi, containerBarcode,
  containerType = "EXPERIMENT_CONTAINER", fullMetadata = FALSE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{Experiment container barcode}

\item{containerType}{Experiment container entity type}

\item{fullMetadata}{- Get full metadata, default is FALSE}

\item{useVerbose}{Boolean on whether to use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{getExperimentContainerContents} - Gets information about experiment container contents.
}
\details{
\code{getExperimentContainerContents} Gets experiment container contents.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
cell <- getExperimentContainerContents(login$coreApi, "BTCR1", "EXPERIMENT_CONTAINTER")
logOut(login$coreApi)
}
}
\author{
Scott Russell scott.russell@thermofisher.com
}
