% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainers.R
\name{getExperimentContainers}
\alias{getExperimentContainers}
\title{getExperimentContainers - Gets experiment containers from experiment identified by barcode.}
\usage{
getExperimentContainers(coreApi, experimentType, experimentBarcode, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type to get}

\item{experimentBarcode}{barcode of experiment to query}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is a list of container barcodes.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getExperimentContainers}  Gets experiment contaniers from experiment identified by experiment barcode.
}
\details{
\code{getExperimentContainers}  Gets experiment containers from experiment identified by barcode.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
exptContainerBarcodes <- getExperimentContainers(login$coreApi, "experimentType", "experimentBarcode")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
