% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentContainersAndExperimentSamples.R
\name{getExperimentContainersAndExperimentSamples}
\alias{getExperimentContainersAndExperimentSamples}
\title{getExperimentContainersAndExperimentSamples - Gets the experiment container, container type and experiment samples of an experiment.
\code{getExperimentContainersAndExperimentSamples}  Gets the experiment container, container type and experiment samples of an experiment.}
\usage{
getExperimentContainersAndExperimentSamples(coreApi, experimentType,
  experimentBarcode, includeExperimentSamples = TRUE,
  fullMetadata = FALSE, ...)
}
\arguments{
\item{experimentType}{entity type which has the attribute metadata}

\item{experimentBarcode}{barcode of entity to get}

\item{includeExperimentSamples}{get experiment sample associated to containers of an experiment, default FALSE}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}

\item{coreAPI}{coreApi object with valid jsessionid}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} list with experiment container, container type and experiment samples of an experiment.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getExperimentContainersAndExperimentSamples} - Gets the experiment container, container type and experiment samples of an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experimentSampleLotNames <- getExperimentContainersAndExperimentSamples(login$coreApi, "experimentType", "EXP123", FALSE)
logOut(login$coreApi)
}
Specific data can be extracted, one example is to extract the container barcodes:
\dontrun{
containerBarcode <- sapply(
    result$entity[[contExpansion]], 
    #contExpansion is "REV_CONTAINER_EXPERIMENT_EXPERIMENT_CONTAINER" or "EXPERIMENT_CONTAINERS" depending on the PFS version.
    FUN = function(x) {
        x$Barcode})
}
}
\author{
Natasha Mora natasha.mora@thermofisher.com
}
