% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleAssayData.R
\name{getExperimentSampleAssayData}
\alias{getExperimentSampleAssayData}
\title{getExperimentSampleAssayData - Gets assay data for an experiment sample.}
\usage{
getExperimentSampleAssayData(coreApi, experimentAssayType,
  experimentSampleBarcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentAssayType}{assay type to get}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{ getExperimentSampleAssayData }  Gets experiment samples from experiment identified by experiment barcode.
Does not retieve files attached as data. Use getExperimentSampleAssayFileData to retrieve assay data that is a file.
}
\details{
\code{getExperimentSampleAssayData }  Gets assay data for a experiment sample identified by barcode.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experiment <- getExperimentSampleAssayData(login$coreApi, "experimentAssayType", "experimentSampleBarcode")
logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
