% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleAssayFileData.R
\name{getExperimentSamplesAssayFileData}
\alias{getExperimentSamplesAssayFileData}
\title{getExperimentSamplesAssayFileData - Gets file attached as assay data.}
\usage{
getExperimentSamplesAssayFileData(coreApi, assayType,
  experimentSampleBarcode, attributeName, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type to get}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{attributeName}{Name of the attibute that containts the file data}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http}
}
\value{
returns a list $entity containsbinary object that in the file, $response contains the entire http response
}
\description{
\code{ getExperimentSamplesAssayFileData }  Gets file attached as assay data.
Use getExperimentSamplesAssayData for non-file data.
}
\details{
\code{getExperimentSamplesAssayFileData }  Gets file attached as assay data.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- getExperimentSamplesAssayFileData(login$coreApi, "assayType", "experimentSampleBarcode", "CI_FILE")
writeBin(response$entity, "myfile.png")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com
}
