% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleIntermediateData.R
\name{getExperimentSamplesIntermediateData-deprecated}
\alias{getExperimentSamplesIntermediateData-deprecated}
\title{getExperimentSamplesIntermediateData - Gets intermediate data for an experiment sample.}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment type for sample}

\item{experimentAssayType}{assay type for sample}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{intermediateDataName}{assay  intermediate data name to retrive as configured in the assay.}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http call}
}
\value{
returns a list $entity contains data frame with derived experiment sample barcodes concentration,
        and assay raw data. $response contains the entire http response
}
\description{
\code{getExperimentSamplesIntermediateData}   Gets raw data for a experiment sample identified by barcode.
}
\details{
\code{getExperimentSamplesIntermediateData}  Gets intermediate data for a experiment sample identified by barcode.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- getExperimentSamplesIntermediateData(
  login$coreApi,
  "experimentType",
  "experimentAssayType",
  "intermediateDataName",
  "experimentSampleBarcode"
)
rawdata <- response$entity
logOut(login$coreApi)
}
}
\seealso{
\code{\link{pfsrsdk-deprecated}}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
\keyword{internal}
