% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSamplesRawData.R
\name{getExperimentSamplesRawData}
\alias{getExperimentSamplesRawData}
\title{getExperimentSamplesRawData - Gets raw data for an experiment container.}
\usage{
getExperimentSamplesRawData(coreApi, experimentContainerBarcode,
  fullMetadata = FALSE, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentContainerBarcode}{experiment sample container of entity to get}

\item{fullMetadata}{- get full metadata, default is FALSE}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http call}
}
\value{
returns a list $entity contains data frame with derived experiment sample barcodes concentration,
        and assay raw data. $response contains the entire http response
}
\description{
\code{ getExperimentSamplesRawData }   Gets raw data for an experiment container identified by barcode.
}
\details{
\code{getExperimentSamplesRawData }  Gets raw data for a experiment container identified by barcode.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- getExperimentSamplesRawData(login$coreApi,
  "experimentContainerBarcode",
  useVerbose = FALSE
)
rawdata <- response$entity
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
