% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSamplesWithSampleLots.R
\name{getExperimentSamplesWithSampleLots}
\alias{getExperimentSamplesWithSampleLots}
\title{getExperimentSamplesWithSampleLots - Gets sample lots assigned to experiment samples of an experiment.
\code{getExperimentSamplesWithSampleLots}  Gets sample lots assigned to experiment samples of an experiment.}
\usage{
getExperimentSamplesWithSampleLots(coreApi, experimentType,
  experimentBarcode, fullMetadata = FALSE, ...)
}
\arguments{
\item{experimentType}{entity type which has the attribute metadata}

\item{experimentBarcode}{barcode of entity to get}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}

\item{coreAPI}{coreApi object with valid jsessionid}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} list with sample lot names for the corresponding experiment sample barcode from a specific experiment.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getExperimentSamplesWithSampleLots} - Gets sample lots assigned to experiment samples of an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experimentSampleLotNames <- getExperimentSamplesWithSampleLots(login$coreApi, "experimentType", "EXP123")
logOut(login$coreApi)
}
Specific data can be extracted, one example is to extract the sample lot names:
\dontrun{
sampleLotName <- sapply(
    response$content[[expansion]],
    FUN = function(x) {
        x$ENTITY$Name})
}
}
\author{
Natasha Mora natasha.mora@thermofisher.com

Edgardo Gutierrez edgardo.gutierrez@thermofisher.com
}
