% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWellContents.R
\name{getWellContents}
\alias{getWellContents}
\title{getWellContents -  Gets content information of a single container well.}
\usage{
getWellContents(coreApi, containerBarcode, containerWellNum,
  containerType = "CONTAINER", fullMetadata = FALSE,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{container barcode}

\item{containerWellNum}{number location of container's well}

\item{containerType}{entity type of container (default: "CONTAINER")}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{useVerbose}{Use verbose communication for debugging (default: FALSE)}
}
\value{
RETURN returns a list $entity contains well information, $response contains the entire http response
}
\description{
\code{getWellContents} - Gets content information of a single container well.
}
\details{
\code{getWellContents} Gets content information of a single container well.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
well <- getWellContents(login$coreApi, "VIA9", "1", "VIAL")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com
}
