% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isExperimentPublished.R
\name{isExperimentPublished}
\alias{isExperimentPublished}
\title{isExperimentPublished gets a boolean indicating if an experiment is published.}
\usage{
isExperimentPublished(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment.}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $status contains the experiment published status, $response contains the entire http response
}
\description{
\code{isExperimentPublished} - gets a boolean indicating if an experiment is published.
}
\details{
\code{isExperimentPublished} gets a boolean indicating if an experiment is published.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
update <- isExperimentPublished(login$coreApi, experimentType, experimentBarcode, useVerbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Natasha Mora, natasha.mora@thermofisher.com
}
