% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadXmlConfigFile.R
\name{loadXmlConfigFile}
\alias{loadXmlConfigFile}
\title{loadXMLconfigFile- Loads an XML config file}
\usage{
loadXmlConfigFile(coreApi, xmlFilePath = NULL, appName = NULL)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{xmlFilePath}{Path to XML file}

\item{appName}{name of PFS app to enable}
}
\value{
RETURN returns the entire http response
}
\description{
\code{xmlFilePath} Loads an XML config file.
}
\details{
\code{xmlFilePathe}  Loads an XML config file
}
\examples{
\dontrun{
api<-CoreAPIV2::coreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
loadXMLconfigFile(coreApi,"PathtoFile/test.xml")
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com

Scott Russell scott.russell@thermofisher.com
}
