% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataCleanName.R
\name{ODATAcleanName-deprecated}
\alias{ODATAcleanName-deprecated}
\title{ODATA clean Name}
\arguments{
\item{name}{string to clean}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{ODATAcleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.
}
\details{
ODATAcleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.

\code{ODATAcleanName} Clean a name for ODATA.
}
\examples{
\dontrun{
new_name <-CoreAPIV2::ODATAcleanName("384 Well Plate")
new_name
_384_WELL_PLATE
 }
}
\seealso{
\code{\link{CoreAPIV2-deprecated}}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com
}
\keyword{internal}
