% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExperimentSampleAssayFileData.R
\name{setExperimentSampleAssayFileData}
\alias{setExperimentSampleAssayFileData}
\title{setExperimentSampleAssayFileData - Puts file attached as assay data into an experiment sample.}
\usage{
setExperimentSampleAssayFileData(coreApi, assayType,
  experimentSampleBarcode, attributeName, filePath, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type that contains sample}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{attributeName}{Name of the attribute that containts the file data}

\item{filePath}{path to file to upload}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http calls}
}
\value{
returns a list $entity contains binary object that in the file, $response contains
the entire http response
}
\description{
\code{ setExperimentSampleAssayFileData } Puts file attached as assay data into an experiment sample.
}
\details{
\code{setExperimentSampleAssayFileData } Puts file attached as assay data into an experiment sample.
}
\examples{
\dontrun{
api <- CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login <- CoreAPIV2::authBasic(api)
response <- setExperimentSampleAssayFileData(login$coreApi,
  "TURBIDITY ASSAY",
  "TBES1",
  "CI_EXTRA_DATA",
  "/path/to/data.file")
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
