% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExperimentSampleAssayFileData.R
\name{setExperimentSamplesAssayFileData-deprecated}
\alias{setExperimentSamplesAssayFileData-deprecated}
\title{setExperimentSamplesAssayFileData - puts file attached as assay data in an experiment.}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{assayType}{assay type that contains sample}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{attributeName}{Name of the attribute that containts the file data}

\item{filePath}{path to file to upload}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http calls}
}
\value{
returns a list $entity contains binary object that in the file, $response contains
the entire http response
}
\description{
\code{ setExperimentSamplesAssayFileData } Puts file attached as assay data
 in an experiment
}
\details{
\code{setExperimentSamplesAssayFileData }  puts file attached as assay data.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- setExperimentSamplesAssayFileData(
  login$coreApi,
  "assayType",
  "barcode",
  "filePath",
  "CI_FILE"
)
logOut(login$coreApi)
}
}
\seealso{
\code{\link{pfsrsdk-deprecated}}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
\keyword{internal}
