% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSemVer.R
\name{setSemVer}
\alias{setSemVer}
\title{setSemVer - Retrieves the PFS SemVer and creates a warning with the right value if null.}
\usage{
setSemVer(coreApi)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}
}
\value{
RETURN returns semVer containing the semVer of the PFS system.
}
\description{
\code{setSemVer} Retrieves the PFS SemVer and creates a warning with the right value if null.
}
\details{
\code{setSemVer} Retrieves the PFS SemVer and creates a warning with the right value if null.
}
\examples{
\dontrun{
api<-CoreAPIV2("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
semver<-CoreAPIV2::setSemVer(login$coreApi)
logOut(login$coreApi )
}
}
\author{
Adam Wheeler, adam.j.wheeler@accenture.com

Natasha Mora natasha.mora@thermofisher.com
}
