% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWellContents.R
\name{setWellContents}
\alias{setWellContents}
\title{setWellContents -  Puts a sample lot in a container well.}
\usage{
setWellContents(coreApi, containerType, containerBarcode, containerWellNum,
  sampleLotType, sampleLotBarcode, amount, amountUnit, concentration,
  concentrationUnit, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container type}

\item{containerBarcode}{barcode of a container that IS NOT assigned to an experiment.}

\item{containerWellNum}{container well number (The plate must have a sample in well A1 in order to setWellContents in other wells.)}

\item{sampleLotType}{sample lot type}

\item{sampleLotBarcode}{barcode of lot to add to well}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{useVerbose}{use verbose communications for debugging}
}
\value{
RETURN returns a list $entity contains updated container
        information, $response contains the entire http response
}
\description{
\code{setWellContents} - Puts a cell lot in a container well.
}
\details{
\code{setWellContents} Puts a sample lot in a container well.
}
\examples{
\dontrun{
api<-CoreAPIV2::CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
well<- setWellContents<-(coreApi, containerType = "CONTAINER", containerBarcode = "96W101", containerWellNum = "1",
                           sampleLotType = "BEER_SAMPLE_LOT", sampleLotBarcode = "BS1000-1", amount = 1, amountUnit = "ml", concentration = 1,
                           concentrationUnit = "nM",useVerbose = FALSE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
