% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCellContents.R
\name{updateCellContents-deprecated}
\alias{updateCellContents-deprecated}
\title{updateCellContents -  Updates amount and concentrations of existing cell.}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container entity type}

\item{containerBarcode}{container barcode}

\item{containerCellNum}{container cell number}

\item{sampleLotBarcode}{barcode of sample lot to add to cell}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{useVerbose}{use verbose communications for debugging}
}
\value{
RETURN returns a list $entity contains updated container information,
  $response contains the entire http response
}
\description{
\code{updateCellContents} - Updates amount and concentrations of existing cell.
}
\details{
\code{updateCellContents} Updates amount and concentrations of existing cell.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- updateCellContents(
  login$coreApi,
  "96 WELL PLATE",
  "96W78",
  "1",
  "SMSA7-1",
  "10.1",
  "mL",
  "0.5",
  "nM"
)
logOut(login$coreApi)
}
}
\seealso{
\code{\link{pfsrsdk-deprecated}}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
\keyword{internal}
