% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntityAttributes.R
\name{updateEntityAttributes}
\alias{updateEntityAttributes}
\title{\code{updateEntityAttributes} Update entity assoications.}
\usage{
updateEntityAttributes(coreApi, entityType, barcode, updateValues,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{updateValues}{vaules to update as list of value pairs}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateEntityAttributes}  Update entity attributes.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
updateValues <- list(SOURCE_LAB = "My Lab", REQUESTOR = "you")
response <- updateEntityAttributes(login$coreApi, "entityType", "barcode", updateValues)
updatedEntity <- response$entity
logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com
}
