% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntityLocation.R
\name{updateEntityLocation}
\alias{updateEntityLocation}
\title{\code{updateEntityLocation}- Update entity location}
\usage{
updateEntityLocation(coreApi, entityType, barcode, locationBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{type of entity to update}

\item{barcode}{barcode of entity to update}

\item{locationBarcode}{barcode of new location}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateEntityLocation} - Update entity location
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- updateEntityLocation(login$coreApi, "entityType", "barcode", "locationBarcode")
entity <- response$entity
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
