% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperimentSampleRawData.R
\name{updateExperimentSampleRawData}
\alias{updateExperimentSampleRawData}
\title{updateExperimentSampleRawData - Update experiment sample raw data.}
\usage{
updateExperimentSampleRawData(coreApi, experimentContainerBarcode,
  rawDataCellNum, rawDataValues, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentContainerBarcode}{User provided barcode as a character string}

\item{rawDataCellNum}{cell (well) number of container}

\item{rawDataValues}{assay attributes as a list of key-values pairs}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information,
       $response contains the entire http response
}
\description{
\code{updateExperimentSampleRawData} Update experiment sample assay raw data.
}
\details{
\code{updateExperimentSampleRawData} Update experiment sample raw data.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- updateExperimentSampleRawData(login$coreApi,
  experimentContainerBarcode = "BTCR1", rawDataCellNum = 1,
  rawDataValues = list(DATA_VALUE = 100, CI_ACCEPT = FALSE)
)

updatedEntity <- response$entity
logOut(login$coreApi)
response <- authBasic(coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
