% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMetadata.R
\name{updateMetadata}
\alias{updateMetadata}
\title{\code{updateMetadata} g Updates cached metadata so metadata is up to date.}
\usage{
updateMetadata(coreApi, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of XML-formatted OData metadata for all entities.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{updateMetadata}  Updates cached metadata so metadata is up to date.
Must be run after any configuration changes.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
metadata <- updateMetadata(login$coreApi, useverbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Scott Russell scott.russell@thermofisher.com
}
