% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiGET.R
\name{apiGET}
\alias{apiGET}
\title{apiGET - Do a GET from the PFS API.}
\usage{
apiGET(coreApi, resource, query, headers = NULL, special = NULL,
  useVerbose = FALSE, useRaw = FALSE, fullReturn = TRUE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{entity type to get}

\item{query}{query string}

\item{headers}{headers to be added to get}

\item{special}{passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging (default FALSE)}

\item{useRaw}{Return raw response content (default FALSE)}

\item{fullReturn}{Return the entire response object, or just the response content (default TRUE)}
}
\value{
List of length 2, containing \code{content} and \code{response} objects:
\itemize{
 \item{\code{content}} is the HTTP response content (concatenated if a chunked response).
 \item{\code{response}} is the entire HTTP response. If chuncked, \code{response} object contains
the full HTTP response of the last HTTP GET call. NULL if \code{fullReturn} is FALSE.
}
}
\description{
\code{apiGET} - Do a GET from the PFS API.
}
\details{
\code{apiGET} Do a GET from the PFS API.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- apiGET(login$coreApi, "json", resource, query)
content <- response$content
error <- response$error$message
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Francisco Marin francisco.marin@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
