% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiPOST.R
\name{apiPOST}
\alias{apiPOST}
\title{apiPOST - Do a POST to the PFS API.}
\usage{
apiPOST(coreApi, resource = NULL, body = NULL, encode,
  headers = NULL, special = NULL, useVerbose = FALSE,
  fullReturn = TRUE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{entity type for POST}

\item{body}{body for request}

\item{encode}{encode type must be "multipart", "form", "json", "raw"}

\item{headers}{headers to be added to \code{httr::POST}.}

\item{special}{passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}

\item{fullReturn}{Return the entire response object, or just the response content (default TRUE)}
}
\value{
List of length 2, containing \code{content} and \code{response} objects:
\itemize{
 \item{\code{content}} is the HTTP response content.
 \item{\code{response}} is the entire HTTP response. NULL if \code{fullReturn} is FALSE.
}
}
\description{
\code{apiPOST} - Do a POST to the PFS API.
}
\details{
\code{apiPOST}  Do a POST to the PFS API.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- apiPOST(login$coreApi, "SAMPLE", body, "json", special = NULL, useVerbose = FALSE)
content <- response$content
error <- response$error$message
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Francisco Marin francisco.marin@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
