% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiPUT.R
\name{apiPUT}
\alias{apiPUT}
\title{apiPUT - Do a PUT to the Core ODATA REST API.}
\usage{
apiPUT(coreApi, resource = NULL, query = NULL, body = NULL, encode,
  headers = NULL, special = NULL, useVerbose = FALSE, unbox = TRUE,
  valueFlag = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{entity type for PUT}

\item{query}{query string}

\item{body}{body for request}

\item{encode}{encoding to use for request option are "multipart", "form", "json", "raw"}

\item{headers}{headers to be added to get.}

\item{special}{passed to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}

\item{unbox}{use autounbox when doing lait yo json conversion}

\item{valueFlag}{Tells the PUT if there needs to be a /$value added to the end.}
}
\value{
Returns the entire http response
}
\description{
\code{apiPUT} - Base PUT call to Core ODATA REST API.
}
\details{
\code{apiPUT}  Do a PUT to the Core ODATA REST API.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- apiPUT(login$coreApi,
  "SAMPLE",
  body,
  "json",
  special = NULL,
  useVerbose = FALSE,
  unbox = TRUE
)
content <- httr::content(response)
logOut(login$coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler, adam.j.wheeler@accenture.com
}
