% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachFile.R
\name{attachFile}
\alias{attachFile}
\title{attachFile - Attaches a file to an entity or file attribute.}
\usage{
attachFile(coreApi, entityType, barcode, filePath,
  targetAttributeName = "", ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{PFS entity type}

\item{barcode}{User provided barcode as a character string}

\item{filePath}{path to the file to attach}

\item{targetAttributeName}{- if included the name if the attribute to attach the file to.  Must be in all caps.}

\item{...}{additional arguments passed to \code{apiPUT}}
}
\value{
List of length 2, containing \code{content} and \code{response} objects:
\itemize{
 \item{\code{content}} is the HTTP response content. If \code{targetAttributeName}
 is not empty, \code{content} will return as NULL.
 \item{\code{response}} is the entire HTTP response. If \code{fullReturn} is FALSE
 will return as NULL.
}
}
\description{
\code{attachFile} Attaches a file to entity identified by barcode.
Note: This function uses the JSON API to post a file to an entity and
Odata to post to an attribute.
}
\details{
\code{attachFile}  Attaches a file to an entity or file attribute.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
modifiedItem <- attachFile(login$coreApi,
  entitytype,
  barcode,
  filePath,
  targetAttributeName = "",
  useVerbose = FALSE
)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
