% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachFile.R
\name{attachFile}
\alias{attachFile}
\title{attachFile - Attaches a file to an entity or file attribute.}
\usage{
attachFile(coreApi, entityType, barcode, filePath,
  targetAttributeName = "", useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{barcode}{User provided barcode as a character string}

\item{targetAttributeName}{- if included the name if the attribute to attach the file to.  Must be in all caps.}

\item{useVerbose}{Use verbose communication for debugging}

\item{filename}{name to use for the attached file}

\item{filepath}{path to the file to attach}
}
\value{
RETURN returns a list $entity contains entity information (if present), $response contains the entire http response
}
\description{
\code{attachFile} Attaches a file to entity identified by barcode.
Note: This function uses the JSON API to post a file to an entity and
Odata to post to an attribute.
}
\details{
\code{attachFile}  Attaches a file to an entity or file attribute.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
modifiedItem<-CoreAPIV2::attachFile(response$coreApi,barcode,filename,
         filepath,targetAttributeName="",useVerbose=FALSE)
CoreAPIV2::logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler, adam.j.wheeler@accenture.com

Scott Russell, scott.russell@thermofisher.com
}
