% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authBasic.R
\name{authBasic}
\alias{authBasic}
\title{authBasic - Authenticates against the ODATA REST API using basic authentication.}
\usage{
authBasic(coreApi, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{object of class coreApi that contains username, password,  baseURL and
tenant. tenant is required if user has access to multiple tenants.}

\item{useVerbose}{- Use verbose settings for HTTP commands}
}
\value{
returns a list with two oblects. coreApi which returns the passed coreApi object with  jsessionid,
           awselb and employeeid populated, $response contains the entire http response
}
\description{
\code{authBasic} Logs in and returns a fully populated coreApi object in $coreAPI.
}
\details{
\code{authBasic} Authenticates against the ODATA REST API using basic authentication.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
response <- authBasic(api)
login <- response$core$Api
error <- httr::http_error(response$response)
logOut(response$coreApi, useVerbose = TRUE)
}
}
\author{
Craig Parman ngsAnalytics, ngsanaltics.com

Scott Russell scott.russell@thermofisher.com

Adam Wheeler, adam.j.wheeler@accenture.com

Natasha Mora natasha.mora@thermofisher.com
}
