% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{coreAPI Creates a object of class coreAPI that contains user and connection information.}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with tenant information in json format.}
}
\value{
Object of class coreAPI
}
\description{
coreAPI Creates a object of class coreAPI that contains user and connection information.
}
\details{
{ Creates a object of class coreAPI that contains username,
         password, host, tenant, port.
         It has slots for tenant, jsessionId, AWSELB, and host.
         Requires a json file that is a POSTMAN environment file.
        \code{#'Creates a object of class coreAPI that contains tenant information}
        \code{coreAPI("path to json")}.
}
The json must include the fields shown below.
    \code{
{
  "values": [
    {
      "key": "tenant",
      "value": "FULLNAMEOFTENANT",
      "type": "text",
      "enabled": true
    },
    {
      "key": "alias",
      "value": "SHORTNAMEINURL",
      "type": "text",
      "enabled": true
    },
    {
      "key": "scheme",
      "value": "https",
      "type": "text",
      "name": "scheme",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "host",
      "value": "HOSTNAME",
      "type": "text",
      "name": "host",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "port",
      "value": "443",
      "type": "text",
      "enabled": true
    },
    {
      "key": "context",
      "value": "",
      "type": "text",
      "enabled": true
    },
    {
      "key": "username",
      "value": "USER",
      "type": "text",
      "name": "nameadmin",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "password",
      "value": "PASSWORD",
      "type": "text",
      "name": "passwordadmin",
      "enabled": true,
      "hovered": false
    },
    {
      "key": "semver",
      "value": "",
      "type": "text",
      "enabled": true
    }
    ]
}
    }

 The tenant value may be set to "" if the user only has access to one tenant.
As an alternative the environment json object from Postman can be used directly.
}
\examples{
\dontrun{
api<-CoreAPIV2::coreAPI("/home/environment.json")
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Scott Russell scott.russell@thermofisher.com

Adam Wheeler, adam.j.wheeler@accenture.com

Natasha Mora natasha.mora@thermofisher.com
}
