% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentUnpublish.R
\name{experimentUnpublish}
\alias{experimentUnpublish}
\title{experimentUnpublish Unpublishes an experiment.}
\usage{
experimentUnpublish(coreApi, experimentType, experimentBarcode, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment (Experiments that require a signature cannot be published through the API).}

\item{...}{additional arguments passed to \code{apiPOST}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content.
  For PFS version 5.x.x the \code{entity} value comes from \code{response$content$response$data}
  For PFS version 6.x.x the \code{entity} value comes from \code{response$content}
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{experimentUnpublish} - Unpublishes an experiment.
}
\details{
\code{experimentUnpublish} Unpublishes an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
update <- experimentUnpublish(login$coreApi, experimentType, experimentBarcode, useVerbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
