% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttachedAttributeFile.R
\name{getAttachedAttributeFile}
\alias{getAttachedAttributeFile}
\title{getAttachedAttributeFile - Gets file attached to an attribute on a entity}
\usage{
getAttachedAttributeFile(coreApi, entityType, barcode, attribute,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type where sample is attached}

\item{barcode}{barcode of the entity}

\item{attribute}{name of the attribute}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used}
}
\value{
returns a list $entity contains the binary file data, $response contains the entire http response
}
\description{
\code{ getAttachedAttributeFile }  Gets file attached to an attribute on a entity.
}
\details{
\code{getAttachedAttributeFile}  Gets file attached to an attribute on a entity
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- getAttachedAttributeFile(login$coreApi, entityType, barcode, attribute)
witeBin(response$entity, "filename.txt")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com
}
