% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityAssociations.R
\name{getEntityAssociations}
\alias{getEntityAssociations}
\title{\code{getEntityAssociations} Get assoication for a context}
\usage{
getEntityAssociations(coreApi, entityType, barcode, associationContext,
  fullMetadata = FALSE, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{associationContext}{association context}

\item{fullMetadata}{- get full metadata, default is FALSE}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http}
}
\value{
returns a list $entity contains entity associations, $response contains the entire http response
}
\description{
\code{getEntityAssociations}  Get assoication for a entity
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
associations <- getEntityAssociations(login$coreAPI, "entityType", "barcode", "associationContext")
logOut(login$coreAPI)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com
}
