% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSampleAssayData.R
\name{getExperimentSampleAssayData}
\alias{getExperimentSampleAssayData}
\title{getExperimentSampleAssayData - Gets assay data for an experiment sample.}
\usage{
getExperimentSampleAssayData(coreApi, experimentAssayType,
  experimentSampleBarcode, fullMetadata = FALSE, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentAssayType}{assay type to get}

\item{experimentSampleBarcode}{experiment sample barcode of entity to get}

\item{fullMetadata}{- get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of entity information.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{ getExperimentSampleAssayData }  Gets assay data for a experiment sample identified by barcode.
}
\details{
\code{getExperimentSampleAssayData }  Gets assay data for a experiment sample identified by barcode.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
experiment <- getExperimentSampleAssayData(login$coreApi, "experimentAssayType", "experimentSampleBarcode")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
