% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentWellContents.R
\name{getExperimentWellContents}
\alias{getExperimentWellContents}
\title{getExperimentWellContents -  Gets content information of a single container well in an experiment.}
\usage{
getExperimentWellContents(coreApi, experimentContainerBarcode,
  experimentContainerWellNum,
  experimentContainerType = "EXPERIMENT_CONTAINER",
  fullMetadata = FALSE, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentContainerBarcode}{experiment container barcode}

\item{experimentContainerWellNum}{number location of experiment container's well}

\item{experimentContainerType}{entity type of experiment container (default: "EXPERIMENT_CONTAINER")}

\item{fullMetadata}{get full metadata, default is FALSE}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of well information.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{getExperimentWellContents} - Gets content information of a single container well in an experiment.
}
\details{
\code{getExperimentWellContents} Gets content information of a single container well in an experiment.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
well <- getExperimentWellContents(login$coreApi, "BTCR1", "1", "BITTERNESS_EXPERIEMENT_CONTAINER")
logOut(login$coreApi)
}
}
\author{
Scott Russell scott.russell@thermofisher.com

Natasha Mora natasha.mora@thermofisher.com
}
