% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentWellContents.R
\name{getExperimentWellContents}
\alias{getExperimentWellContents}
\title{getExperimentWellContents -  Gets content information of a single container well in an experiment.}
\usage{
getExperimentWellContents(coreApi, experimentContainerBarcode,
  experimentContainerWellNum,
  experimentContainerType = "EXPERIMENT_CONTAINER", useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentContainerBarcode}{experiment container barcode}

\item{experimentContainerWellNum}{number location of experiment container's well}

\item{experimentContainerType}{entity type of experiment container (default: "EXPERIMENT_CONTAINER")}

\item{useVerbose}{Use verbose communication for debugging (default: FALSE)}
}
\value{
RETURN returns a list $entity contains well information, $response contains the entire http response
}
\description{
\code{getExperimentWellContents} - Gets content information of a single container well in an experiment.
}
\details{
\code{getExperimentWellContents} Gets content information of a single container well in an experiment.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPIV2::authBasic(api)
well<-CoreAPIV2::getExperimentWellContents(login$coreApi,"BTCR1","1", "BITTERNESS_EXPERIEMENT_CONTAINER")
CoreAPIV2::logOut(login$coreApi)
}
}
\author{
Scott Russell scott.russell@thermofisher.com
}
