% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isExperimentPublished.R
\name{isExperimentPublished}
\alias{isExperimentPublished}
\title{isExperimentPublished gets a boolean indicating if an experiment is published.}
\usage{
isExperimentPublished(coreApi, experimentType, experimentBarcode, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment.}

\item{...}{additional arguments passed to \code{apiGET}}
}
\value{
List of length 2, containing \code{content} and \code{response} objects:
\itemize{
 \item{\code{content}} is the HTTP response content of experiment published status.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{isExperimentPublished} - gets a boolean indicating if an experiment is published.
}
\details{
\code{isExperimentPublished} gets a boolean indicating if an experiment is published.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
update <- isExperimentPublished(login$coreApi, experimentType, experimentBarcode, useVerbose = TRUE)
logOut(login$coreApi)
}
}
\author{
Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
