% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadXmlConfigFile.R
\name{loadXmlConfigFile}
\alias{loadXmlConfigFile}
\title{loadXMLconfigFile- Loads an XML config file}
\usage{
loadXmlConfigFile(coreApi, xmlFilePath = NULL, appName = NULL)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{xmlFilePath}{Path to XML file}

\item{appName}{name of PFS app to enable}
}
\value{
RETURN returns the entire http response
}
\description{
\code{xmlFilePath} Loads an XML config file.
}
\details{
\code{xmlFilePathe}  Loads an XML config file
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
loadXMLconfigFile(login$coreApi, "PathtoFile/test.xml")
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
