% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataCleanName.R
\name{odataCleanName}
\alias{odataCleanName}
\title{odataCleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.}
\usage{
odataCleanName(name, refType = "odataObject")
}
\arguments{
\item{name}{string to clean}

\item{refType}{Reference to the type of object being passed (default: odataObject).}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{odataCleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.
}
\details{
\code{odataCleanName} Clean a name for ODATA.
}
\section{Using \code{refType} parameter}{

\code{refType} accepts two values, the default "odataObject" and "tenant".
"odataObject" ensures that a leading underscore is placed in front of numbers
at the beginning of the string to comply with OData Standards. This function
can be used for other objects as well to replace spaces and hyphens with
underscores. When \code{refType} is set to "tenant", underscores are
substituted for spaces, but leaves hyphens intact.
}

\examples{
\dontrun{
new_name <- odataCleanName("384 Well Plate")
# returns "_384_WELL_PLATE"
new_name <- odataCleanName("384 Well Plate", "tenant")
# returns "384_WELL_PLATE"
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
