% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataCleanName.R
\name{odataCleanName}
\alias{odataCleanName}
\title{odataCleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.}
\usage{
odataCleanName(name)
}
\arguments{
\item{name}{string to clean}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{odataCleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.
}
\details{
\code{odataCleanName} Clean a name for ODATA.
}
\examples{
\dontrun{
new_name <-CoreAPIV2::odataCleanName("384 Well Plate")
new_name
_384_WELL_PLATE
 }
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com
}
