% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odataCleanName.R
\name{odataCleanName}
\alias{odataCleanName}
\title{odataCleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.}
\usage{
odataCleanName(name, refType = "odataObject")
}
\arguments{
\item{name}{string to clean}

\item{refType}{Reference to the type of object being passed. Defaults to odataObject passing another value will prevent the leading underscores from being added to numbers}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{odataCleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.
}
\details{
\code{odataCleanName} Clean a name for ODATA.
}
\examples{
\dontrun{
new_name <- odataCleanName("384 Well Plate")
>_384_Well_Plate
new_name <- odataCleanName("384 Well Plate","tenent")
>384_Well_Plate
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com

Scott Russell scott.russell@thermofisher.com
}
