% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttachedAttributeFile.R,
%   R/getExperimentSampleIntermediateData.R, R/odataCleanName.R,
%   R/pfsrsdk-deprecated.r, R/setExperimentSampleAssayFileData.R,
%   R/updateCellContents.R
\name{getAttachedFile}
\alias{getAttachedFile}
\alias{getExperimentSamplesIntermediateData}
\alias{pfsrsdk-deprecated}
\alias{ODATAcleanName}
\alias{setExperimentSamplesAssayFileData}
\alias{updateCellContents}
\title{ODATAcleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.}
\usage{
getAttachedFile(coreApi, entityType, barcode, attribute,
  useVerbose = FALSE)

getExperimentSamplesIntermediateData(coreApi, experimentType,
  experimentAssayType, intermediateDataName, experimentSampleBarcode,
  useVerbose = FALSE)

ODATAcleanName(name)

setExperimentSamplesAssayFileData(coreApi, assayType,
  experimentSampleBarcode, attributeName, filePath, useVerbose = FALSE)

updateCellContents(coreApi, containerType, containerBarcode,
  containerCellNum, sampleLotBarcode, amount, amountUnit, concentration,
  concentrationUnit, useVerbose = FALSE)
}
\arguments{
\item{name}{string to clean}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{ODATAcleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.

The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\details{
\code{ODATAcleanName} Clean a name for ODATA.
}
\section{\code{getAttachedFile}}{

For \code{getAttachedFile}, use \code{\link{getAttachedAttributeFile}}.
}

\section{\code{getExperimentSamplesIntermediateData}}{

For \code{getExperimentSamplesIntermediateData}, use \code{\link{getExperimentSampleIntermediateData}}.
}

\section{\code{ODATAcleanName}}{

For \code{ODATAcleanName}, use \code{\link{odataCleanName}}.
}

\section{\code{setExperimentSamplesAssayFileData}}{

For \code{setExperimentSamplesAssayFileData}, use \code{\link{setExperimentSampleAssayFileData}}.
}

\section{\code{updateCellContents}}{

For \code{updateCellContents}, use \code{\link{setCellContents}}.
}

\examples{
\dontrun{
new_name <- ODATAcleanName("384 Well Plate")
}
}
\seealso{
\code{\link{pfsrsdk-deprecated}}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Adam Wheeler adam.j.wheeler@accenture.com
}
\keyword{internal}
