% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCellContents.R
\name{setCellContents}
\alias{setCellContents}
\title{setCellContents -  Puts a sample lot in a container Cell.}
\usage{
setCellContents(coreApi, containerType, containerBarcode, containerCellId,
  sampleLotType, sampleLotBarcode, amount, amountUnit, concentration,
  concentrationUnit, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container type}

\item{containerBarcode}{barcode of a container that IS NOT assigned to an experiment.}

\item{containerCellId}{container cell id. (In PFS 5.3.8 (semVer 2.7.1) if multi-wells are used, cell A1 has to be filled or filled first in order to setCellContents in other cells.)}

\item{sampleLotType}{sample lot type}

\item{sampleLotBarcode}{barcode of lot to add to cell}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{...}{additional arguments passed to \code{apiPOST}}
}
\value{
List of length 2, containing \code{content} and \code{response} objects:
\itemize{
 \item{\code{content}} is the HTTP response content of updated cell information.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{setCellContents} - Puts a cell lot in a container cell.
}
\details{
\code{setCellContents} Puts a sample lot in a container cell.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
cell <- setCellContents(login$coreAPI, containerType, containerBarcode, containerCellId,
  sampleLotType, sampleLotBarcode, amount, amountUnit, concentration,
  concentrationUnit,
  useVerbose = FALSE
)
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
