% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntityAttributes.R
\name{updateEntityAttributes}
\alias{updateEntityAttributes}
\title{\code{updateEntityAttributes} Update entity assoications.}
\usage{
updateEntityAttributes(coreApi, entityType, barcode, updateValues, ...)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{updateValues}{vaules to update as list of value pairs}

\item{...}{additional arguments passed to \code{apiPUT}}
}
\value{
List of length 2, containing \code{entity} and \code{response} objects:
\itemize{
 \item{\code{entity}} is the HTTP response content of updated entity information.
 \item{\code{response}} is the entire HTTP response.
}
}
\description{
\code{updateEntityAttributes}  Update entity attributes.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
updateValues <- list(SOURCE_LAB = "My Lab", REQUESTOR = "you")
response <- updateEntityAttributes(login$coreApi, "entityType", "barcode", updateValues)
updatedEntity <- response$entity
logOut(login$coreApi)
}
}
\author{
Craig Parman info@ngsanalytics.com

Adam Wheeler adam.wheeler@thermofisher.com

Scott Russell scott.russell@thermofisher.com
}
