% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperimentSampleData.R
\name{updateExperimentSampleData}
\alias{updateExperimentSampleData}
\title{updateExperimentSampleData - Update experiment sample assay data.}
\usage{
updateExperimentSampleData(coreApi, experimentAssayType,
  experimentSampleBarcode, experimentAssayUpdateAttrList,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentAssayType}{entity type to get as character string}

\item{experimentSampleBarcode}{User provided barcode as a character string}

\item{experimentAssayUpdateAttrList}{assay attributes as a list of key-values pairs}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information,
       $response contains the entire http response
}
\description{
\code{updateExperimentSampleData} Update experiment sample assay data.
}
\details{
\code{updateExperimentSampleData} Update experiment sample assay data.
}
\examples{
\dontrun{
api <- coreAPI("PATH TO JSON FILE")
login <- authBasic(api)
response <- updateExperimentSampleData(login$coreApi,
  experimentAssayType = "BITTERNESS_ASSAY",
  experimentSampleBarcode = "BTES3", experimentAssayUpdateAttrList = list(CI_BITTERNESS_IBU = 9.7, CI_ACCEPT = FALSE)
)
updatedEntity <- response$entity
logOut(login$coreApi)
response <- authBasic(coreApi)
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com

Natasha Mora natasha.mora@thermofisher.com
}
